+++ frontend/type2/client/.gitignore
@@ -13,6 +13,7 @@ dist-ssr
 *.local
 
 .env
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
