+++ frontend/type2/client/src/assets/Icons/AddIcon.jsx
