+++ frontend/type2/client/src/assets/Icons/DeleteIcon.jsx
