+++ frontend/type2/client/src/assets/Icons/FavoriteIcon.jsx
