+++ frontend/type2/client/src/assets/Icons/HeartIcon.jsx
