+++ frontend/type2/client/src/assets/Icons/MypageIcon.jsx
