+++ frontend/type2/client/src/assets/Icons/SearchIcon.jsx
