+++ frontend/type2/client/src/assets/Icons/SupplementsIcon.jsx
