+++ frontend/type2/client/src/assets/Icons/ThumbsDownIcon.jsx
