+++ frontend/type2/client/src/assets/Icons/Thumbs_up_icon.jsx
