+++ frontend/type2/client/src/assets/Icons/UpdateIcon.jsx
