+++ frontend/type2/client/src/assets/Images/my_location_ blue.png
