+++ frontend/type2/client/src/assets/Images/mylocation.png
