--- frontend/type2/client/src/page/Favorites.jsx
@@ -1,9 +1,0 @@
-function Favorites(){
-    return (
-        <>
-            즐겨찾기
-        </>
-    )
-}
-
-export default Favorites;
