+++ frontend/type2/client/src/page/NsRecommendation/NsRecommendation.module.css
