+++ frontend/type2/client/src/page/PetRegister/PetRegErrorPopup.jsx
