+++ frontend/type2/client/src/page/Search/SearchResultsSidebar.module.css
